+++ packages/typespec-rust/test/cadlranch/encode/bytes/tests/bytes_property_client_test.rs
@@ -1,0 +1,79 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+//
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+use cadl_bytes::{
+    models::{
+        Base64BytesProperty, Base64urlArrayBytesProperty, Base64urlBytesProperty,
+        DefaultBytesProperty,
+    },
+    BytesClient,
+};
+
+#[async_std::test]
+async fn base64() {
+    let client = BytesClient::with_no_credential("http://localhost:3000", None).unwrap();
+    let mut input = Base64BytesProperty::default();
+    input.value = Some("test".as_bytes().to_owned());
+    let resp = client
+        .get_bytes_property_client()
+        .base64(input.try_into().unwrap(), None)
+        .await
+        .unwrap();
+    let output: Base64BytesProperty = resp.try_into().unwrap();
+    assert_eq!(output.value, Some("test".as_bytes().to_owned()));
+}
+
+#[async_std::test]
+async fn base64_url() {
+    let client = BytesClient::with_no_credential("http://localhost:3000", None).unwrap();
+    let mut input = Base64urlBytesProperty::default();
+    input.value = Some("test".as_bytes().to_owned());
+    let resp = client
+        .get_bytes_property_client()
+        .base64_url(input.try_into().unwrap(), None)
+        .await
+        .unwrap();
+    let output: Base64urlBytesProperty = resp.try_into().unwrap();
+    assert_eq!(output.value, Some("test".as_bytes().to_owned()));
+}
+
+#[async_std::test]
+#[should_panic]
+async fn base64_url_array() {
+    // TODO: https://github.com/Azure/typespec-rust/issues/56
+    // specifically need to handle nested arrays of base64 encoded bytes
+    let client = BytesClient::with_no_credential("http://localhost:3000", None).unwrap();
+    let mut input = Base64urlArrayBytesProperty::default();
+    input.value = Some(vec![
+        "test".as_bytes().to_owned(),
+        "test".as_bytes().to_owned(),
+    ]);
+    let resp = client
+        .get_bytes_property_client()
+        .base64_url_array(input.try_into().unwrap(), None)
+        .await
+        .unwrap();
+    let output: Base64urlArrayBytesProperty = resp.try_into().unwrap();
+    assert_eq!(
+        output.value,
+        Some(vec![
+            "test".as_bytes().to_owned(),
+            "test".as_bytes().to_owned()
+        ])
+    );
+}
+
+#[async_std::test]
+async fn default() {
+    let client = BytesClient::with_no_credential("http://localhost:3000", None).unwrap();
+    let mut input = DefaultBytesProperty::default();
+    input.value = Some("test".as_bytes().to_owned());
+    let resp = client
+        .get_bytes_property_client()
+        .default(input.try_into().unwrap(), None)
+        .await
+        .unwrap();
+    let output: DefaultBytesProperty = resp.try_into().unwrap();
+    assert_eq!(output.value, Some("test".as_bytes().to_owned()));
+}